requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
Django==4.2.23
django-bootstrap-v5==1.0.11
django-jazzmin==3.0.1
django-location-field==2.7.3
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
